pyproject.toml
[tool.poetry]
name = "handsome"
version = "0.1.0"
description = "discord bot for handsome boys"
authors = ["Andrew Herbig <notandrewherbig@gmail.com>"]

[tool.poetry.scripts]
bot = "handsome:main"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git", extras = ["voice"]}
toml = "^0.10.0"
pydantic = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
pre-commit = "^2.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
